wishlist basics done

+++ app/controllers/process_delete_wish.php
@@ -7,13 +7,13 @@ if (isset($_POST['productId'])) {
     $userId = $_SESSION['id'];
     $productId = $_POST['productId'];
 
-    $sql = " SELECT * FROM tbl_wishlists WHERE user_id='$userId' AND product_id=$productId ";
+    $sql = " SELECT * FROM tbl_wishlists WHERE user_id=$userId AND product_id=$productId ";
 	$result = mysqli_query($conn, $sql);
     $count = mysqli_num_rows($result);
 
     if($count) {
         $row = mysqli_fetch_assoc($result);
-        $sql = " DELETE FROM tbl_wishlists WHERE product_id=$productId ";
+        $sql = " DELETE FROM tbl_wishlists WHERE product_id=$productId AND user_id=$userId ";
         $result = mysqli_query($conn, $sql);
 
         echo $result;
